requirements.txt
asgiref==3.11.0
certifi==2025.11.12
charset-normalizer==3.4.4
Django==6.0
django-js-asset==3.1.2
django-mptt==0.18.0
djangorestframework==3.16.1
idna==3.11
iiif-prezi3==3.0.0
pillow==12.0.0
psycopg2-binary==2.9.11
pydantic==1.10.24
python-dotenv==1.2.1
requests==2.32.5
sqlparse==0.5.4
typing_extensions==4.15.0
urllib3==2.6.1
